Add files via upload
I chose to remove the contact page from the website  as I felt it didn't serve much purpose. I also chose to remove the space I was using for a small icon photo or logo and rather have the menu take up the full width of the page. Colours on the page have been subtly changed to hex codes that match colours present in the background image so as to create a more cohesive and aesthetically pleasing page. A favicon has also been added. The home page is now underway with two columns and a button that will allow the user to download a copy of my CV.